feat: Introduce Attachment module for handling file attachments and update related logic
