+++ crates/forge_provider/src/open_router/transformers/drop_tool_call.rs
@@ -59,6 +59,7 @@ mod tests {
             ],
             tools: vec![],
             tool_choice: None,
+            max_tokens: None,
         };
 
         let request = OpenRouterRequest::from(context);
