+++ crates/forge_provider/src/open_router/transformers/set_cache.rs
@@ -39,6 +39,7 @@ mod tests {
             })],
             tools: vec![],
             tool_choice: None,
+            max_tokens: None,
         };
 
         let request =
